2017 - day18 part 1, converted registers to Long

+++ src/cpu/Computer.kt
@@ -1,13 +1,13 @@
 package cpu
 
 class Computer(var programCounter: Int = 0) {
-    val registers = mutableMapOf<String, Int>()
-    fun getRegister(register: String): Int = registers.getOrDefault(register, 0)
-    fun setRegister(register: String, value: Int): Computer {
+    val registers = mutableMapOf<String, Long>()
+    fun getRegister(register: String): Long = registers.getOrDefault(register, 0)
+    fun setRegister(register: String, value: Long): Computer {
         registers[register] = value
         return this
     }
-    fun operateOnRegister(register: String, action: (Int) -> Int): Computer {
+    fun operateOnRegister(register: String, action: (Long) -> Long): Computer {
         registers[register] = action(registers.getOrDefault(register, 0))
         return this
     }
